Set Stochast Duration to double Related to Issue [WTI-1303]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@10469 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: a2dc9cc2e8ad023610e285968bf335d4b6abb322
Former-commit-id: fa33f9a2c1f5eeea97147c8823c9f27345a2ef83

+++ Ringtoets/Common/src/Ringtoets.Common.Data/Hydraulics/IllustrationPoints/RealizedStochast.cs
@@ -38,7 +38,7 @@ namespace Ringtoets.Common.Data.Hydraulics.IllustrationPoints
         /// <param name="realization">The realization of the stochast.</param>
         /// <exception cref="ArgumentNullException">Thrown when <paramref name="name"/>
         /// is <c>null</c>.</exception>
-        public RealizedStochast(string name, int duration, double alpha, double realization)
+        public RealizedStochast(string name, double duration, double alpha, double realization)
             : base(name, duration, alpha)
         {
             Realization = new RoundedDouble(5, realization);
